Fixed bug that defaultHeaders get inadvertently modified leading to possible side effects in subsequent calls (#38)
* Fixed issue that user-defined retry spacing was ignored

* Fixed type incompatibility issue that could cause possible data loss

* Converted tabs to whitespaces to satisfy spotlessJava check :)

* Fixed bug that defaultHeaders get inadvertently modified leading to possible side effects in subsequent calls.

The previous code referenced the defaultHeaders map instead of _copying_ the contents. This means that all headers added by a call would become part of the defaultHeaders and be silently added to all subsequent calls. This can lead to very strange side effects (in my case 401 Unauthorized error due to a spurious "Authorization" header added from an unrelated call!)